Extract timezone related functions
